# api/inventory/v1/resource_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.7.3
// - protoc             v5.27.0
// source: inventory/v1/resource.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationResourceServiceGetResource = "/api.inventory.v1.ResourceService/GetResource"
const OperationResourceServiceListResources = "/api.inventory.v1.ResourceService/ListResources"

type ResourceServiceHTTPServer interface {
	GetResource(context.Context, *GetResourceRequest) (*GetResourceResponse, error)
	ListResources(context.Context, *ListResourcesRequest) (*ListResourcesResponse, error)
}

func RegisterResourceServiceHTTPServer(s *http.Server, srv ResourceServiceHTTPServer) {
	r := s.Route("/")
	r.GET("/api/inventory/v1.0/resources", _ResourceService_ListResources0_HTTP_Handler(srv))
	r.GET("/api/inventory/v1.0/resources/{id}", _ResourceService_GetResource0_HTTP_Handler(srv))
}

func _ResourceService_ListResources0_HTTP_Handler(srv ResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListResourcesRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationResourceServiceListResources)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListResources(ctx, req.(*ListResourcesRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListResourcesResponse)
		return ctx.Result(200, reply)
	}
}

func _ResourceService_GetResource0_HTTP_Handler(srv ResourceServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetResourceRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationResourceServiceGetResource)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetResource(ctx, req.(*GetResourceRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetResourceResponse)
		return ctx.Result(200, reply)
	}
}

type ResourceServiceHTTPClient interface {
	GetResource(ctx context.Context, req *GetResourceRequest, opts ...http.CallOption) (rsp *GetResourceResponse, err error)
	ListResources(ctx context.Context, req *ListResourcesRequest, opts ...http.CallOption) (rsp *ListResourcesResponse, err error)
}

type ResourceServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewResourceServiceHTTPClient(client *http.Client) ResourceServiceHTTPClient {
	return &ResourceServiceHTTPClientImpl{client}
}

func (c *ResourceServiceHTTPClientImpl) GetResource(ctx context.Context, in *GetResourceRequest, opts ...http.CallOption) (*GetResourceResponse, error) {
	var out GetResourceResponse
	pattern := "/api/inventory/v1.0/resources/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationResourceServiceGetResource))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ResourceServiceHTTPClientImpl) ListResources(ctx context.Context, in *ListResourcesRequest, opts ...http.CallOption) (*ListResourcesResponse, error) {
	var out ListResourcesResponse
	pattern := "/api/inventory/v1.0/resources"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationResourceServiceListResources))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}