# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/inventory/v1.0/resources:
        get:
            tags:
                - ResourceService
            operationId: ResourceService_ListResources
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inventory.v1.ListResourcesResponse'
    /api/inventory/v1.0/resources/{id}:
        get:
            tags:
                - ResourceService
            operationId: ResourceService_GetResource
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.inventory.v1.GetResourceResponse'
components:
    schemas:
...
        api.inventory.v1.GetResourceResponse:
            type: object
            properties:
                resource:
                    $ref: '#/components/schemas/api.inventory.v1.ResourceOut'
            description: The GetResourceResponse message
...
        api.inventory.v1.ListResourcesResponse:
            type: object
            properties:
                response:
                    $ref: '#/components/schemas/api.inventory.v1.PagedResponse'
            description: The ListResourcesResponse message
        api.inventory.v1.PagedResponse:
            type: object
            properties:
                metadata:
                    $ref: '#/components/schemas/api.inventory.v1.PagedResponseMetadata'
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inventory.v1.ResourceOut'
            description: The PagedResponse message (generic)
        api.inventory.v1.PagedResponseMetadata:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                size:
                    type: integer
                    format: int32
                total:
                    type: string
            description: The PagedResponseMetadata message
        api.inventory.v1.Reporter:
            type: object
            properties:
                name:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                lastUpdatedAt:
                    type: string
                    format: date-time
                type:
                    type: string
                url:
                    type: string
                consoleHref:
                    type: string
                apiHref:
                    type: string
                resourceIdAlias:
                    type: string
            description: The Reporter message
        api.inventory.v1.Resource:
            type: object
            properties:
                id:
                    type: string
                displayName:
                    type: string
                tags:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inventory.v1.ResourceTag'
                resourceType:
                    type: string
                reporters:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.inventory.v1.Reporter'
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
            description: The Resource message
...
        api.inventory.v1.ResourceOut:
            type: object
            properties:
                resource:
                    $ref: '#/components/schemas/api.inventory.v1.Resource'
                href:
                    type: string
            description: The ResourceOut message
        api.inventory.v1.ResourceTag:
            type: object
            properties:
                id:
                    type: string
                resourceId:
                    type: string
                namespace:
                    type: string
                key:
                    type: string
                value:
                    type: string
...